Export moved config definitions.
