+++ configdefinitions/src/main/java/com/yahoo/vespa/config/search/core/package-info.java
