+++ configdefinitions/src/vespa/CMakeLists.txt
@@ -80,3 +80,6 @@ vespa_generate_config(configdefinitions ranking-expressions.def)
 install_config_definition(ranking-expressions.def vespa.config.search.core.ranking-expressions.def)
 vespa_generate_config(configdefinitions onnx-models.def)
 install_config_definition(onnx-models.def vespa.config.search.core.onnx-models.def)
+vespa_generate_config(configdefinitions proton.def)
+install_config_definition(proton.def vespa.config.search.core.proton.def)
+vespa_generate_config(configdefinitions hwinfo.def)
