+++ configdefinitions/src/vespa/hwinfo.def
