+++ searchcore/src/tests/proton/common/hw_info_sampler/CMakeLists.txt
@@ -4,6 +4,5 @@ vespa_add_executable(searchcore_hw_info_sampler_test_app TEST
     hw_info_sampler_test.cpp
     DEPENDS
     searchcore_pcommon
-    searchcore_fconfig
 )
 vespa_add_test(NAME searchcore_hw_info_sampler_test_app COMMAND searchcore_hw_info_sampler_test_app)
