+++ searchcore/src/tests/proton/documentdb/CMakeLists.txt
@@ -19,7 +19,6 @@ vespa_add_executable(searchcore_documentdb_test_app TEST
     searchcore_pcommon
     searchcore_grouping
     searchcore_proton_metrics
-    searchcore_fconfig
 )
 vespa_add_test(NAME searchcore_documentdb_test_app COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/documentdb_test.sh
                DEPENDS searchcore_documentdb_test_app)
