+++ searchcore/src/tests/proton/documentdb/buckethandler/CMakeLists.txt
@@ -12,6 +12,5 @@ vespa_add_executable(searchcore_buckethandler_test_app TEST
     searchcore_bucketdb
     searchcore_pcommon
     searchcore_grouping
-    searchcore_fconfig
 )
 vespa_add_test(NAME searchcore_buckethandler_test_app COMMAND searchcore_buckethandler_test_app)
