+++ searchcore/src/tests/proton/documentdb/clusterstatehandler/CMakeLists.txt
@@ -10,6 +10,5 @@ vespa_add_executable(searchcore_clusterstatehandler_test_app TEST
     searchcore_attribute
     searchcore_pcommon
     searchcore_grouping
-    searchcore_fconfig
 )
 vespa_add_test(NAME searchcore_clusterstatehandler_test_app COMMAND searchcore_clusterstatehandler_test_app)
