+++ searchcore/src/tests/proton/documentdb/combiningfeedview/CMakeLists.txt
@@ -13,6 +13,5 @@ vespa_add_executable(searchcore_combiningfeedview_test_app TEST
     searchcore_pcommon
     searchcore_grouping
     searchcore_proton_metrics
-    searchcore_fconfig
 )
 vespa_add_test(NAME searchcore_combiningfeedview_test_app COMMAND searchcore_combiningfeedview_test_app)
