+++ searchcore/src/tests/proton/documentdb/document_subdbs/CMakeLists.txt
@@ -19,6 +19,5 @@ vespa_add_executable(searchcore_document_subdbs_test_app TEST
     searchcore_pcommon
     searchcore_grouping
     searchcore_proton_metrics
-    searchcore_fconfig
 )
 vespa_add_test(NAME searchcore_document_subdbs_test_app COMMAND searchcore_document_subdbs_test_app)
