+++ searchcore/src/tests/proton/documentdb/feedhandler/CMakeLists.txt
@@ -13,7 +13,6 @@ vespa_add_executable(searchcore_feedhandler_test_app TEST
     searchcore_pcommon
     searchcore_grouping
     searchcore_proton_metrics
-    searchcore_fconfig
 )
 vespa_add_test(NAME searchcore_feedhandler_test_app COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/feedhandler_test.sh
                DEPENDS searchcore_feedhandler_test_app)
