+++ searchcore/src/tests/proton/documentdb/fileconfigmanager/CMakeLists.txt
@@ -6,7 +6,6 @@ vespa_add_executable(searchcore_fileconfigmanager_test_app TEST
     searchcore_test
     searchcore_server
     searchcore_pcommon
-    searchcore_fconfig
 )
 vespa_add_test(NAME searchcore_fileconfigmanager_test_app COMMAND ${CMAKE_CURRENT_SOURCE_DIR}/fileconfigmanager_test.sh
                DEPENDS searchcore_fileconfigmanager_test_app)
