+++ searchcore/src/tests/proton/documentdb/threading_service_config/threading_service_config_test.cpp
@@ -1,6 +1,6 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include <vespa/searchcore/config/config-proton.h>
+#include <vespa/config-proton.h>
 #include <vespa/searchcore/proton/common/hw_info.h>
 #include <vespa/searchcore/proton/server/threading_service_config.h>
 #include <vespa/vespalib/testkit/testapp.h>
