+++ searchcore/src/tests/proton/documentmetastore/CMakeLists.txt
@@ -8,7 +8,6 @@ vespa_add_executable(searchcore_documentmetastore_test_app TEST
     searchcore_bucketdb
     searchcore_attribute
     searchcore_feedoperation
-    searchcore_fconfig
     GTest::GTest
 )
 vespa_add_test(NAME searchcore_documentmetastore_test_app COMMAND searchcore_documentmetastore_test_app)
