+++ searchcore/src/tests/proton/proton_configurer/CMakeLists.txt
@@ -4,7 +4,6 @@ vespa_add_executable(searchcore_proton_configurer_test_app TEST
     proton_configurer_test.cpp
     DEPENDS
     searchcore_server
-    searchcore_fconfig
     GTest::GTest
 )
 vespa_add_test(NAME searchcore_proton_configurer_test_app COMMAND searchcore_proton_configurer_test_app)
