+++ searchcore/src/vespa/searchcore/bmcluster/CMakeLists.txt
@@ -48,7 +48,6 @@ vespa_add_library(searchcore_bmcluster STATIC
     searchcore_pcommon
     searchcore_grouping
     searchcore_proton_metrics
-    searchcore_fconfig
     searchcorespi
     storageserver_storageapp
     messagebus_messagebus-test
