--- searchcore/src/vespa/searchcore/config/.gitignore
@@ -1,4 +1,0 @@
-.depend
-Makefile
-config-*.cpp
-config-*.h
