--- searchcore/src/vespa/searchcore/config/CMakeLists.txt
@@ -1,8 +1,0 @@
-# Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-vespa_add_library(searchcore_fconfig STATIC
-    SOURCES
-    DEPENDS
-)
-vespa_generate_config(searchcore_fconfig proton.def)
-install_config_definition(proton.def vespa.config.search.core.proton.def)
-vespa_generate_config(searchcore_fconfig hwinfo.def)
