--- searchcore/src/vespa/searchcore/config/OWNERS
@@ -1,1 +1,0 @@
-geirst
