+++ searchcore/src/vespa/searchcore/proton/server/prepare_restart_handler.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include <vespa/searchcore/config/config-proton.h>
+#include <vespa/config-proton.h>
 #include <condition_variable>
 #include <memory>
 #include <mutex>
